Step 5 Overhaul
Separated cards into individual files for rendering on suggestions page. Still have to do step 3 and 4
