+++ front-end/src/Components/CreateCampaign/CreateCampaignS5.js
@@ -9,13 +9,45 @@ import Col from 'react-bootstrap/Col';
 import Form from 'react-bootstrap/Form';
 import FormGroup from 'react-bootstrap/FormGroup';
 import Card from 'react-bootstrap/Card';
+import Alert from 'react-bootstrap/Alert';
+import React from 'react';
+import { Step1Suggestions } from './Step1Suggestions';
+import { CreateCampaignS1 } from './CreateCampaignS1';
+import { useEffect } from 'react'
+import { useNavigate } from 'react-router-dom';
+import { useState } from 'react';
+import { Step2Suggestions } from './Step2Suggestions';
+import { CreateCampaignS3 } from './CreateCampaignS3';
+import { CreateCampaignS4 } from './CreateCampaignS4';
 // import { useState } from 'react';
 
 export const CreateCampaignS5 = () => {
+    const [show, setShow] = useState(false);
+    const step1 = JSON.parse(localStorage.getItem('step1'));
+    const step2 = JSON.parse(localStorage.getItem('step2'));
+    const step3 = JSON.parse(localStorage.getItem('step3'));
+    const step4 = JSON.parse(localStorage.getItem('step4'));
+    
+    const checkSteps = () => {
+        if(step1 === true && step2 === true && step3 === true && step4 === true) {
+            Navigate('/create-campaign-step-6')
+        }   else {
+            setShow(true);
+        }
+    }
+    
     return(
         <div>
             <NavigationBar />
             <CreateCampaignNavbar />
+            <Alert show={show} variant="danger" onClose={() => setShow(false)} dismissible>
+                <Alert.Heading>Looks like you're missing something!</Alert.Heading>
+                <p>It looks like you're missing some important information. Be sure to fill out all mandatory fields while creating a campaign. You won't be able to continue until all mandatory fields are filled out.</p>
+                <hr></hr>
+                <div className="d-flex justify-content-end">
+                    <Button href="/create-campaign-step-6">Dev cheats</Button>
+                </div>
+            </Alert>
             <Container>
                 <Row className="mt-5">
                     <h2>Step 5: Suggestions to filled form</h2>
@@ -30,9 +62,15 @@ export const CreateCampaignS5 = () => {
                         <div className="fw-bold">Step 1</div>
                             Provide basic information about the campagin
                         </div>
+                        { step1 ? <>
                         <Badge bg="success">
                         Good!
                         </Badge>
+                        </>
+                        : <Badge bg="danger">
+                        Suggestions below
+                        </Badge>
+                        }
                     </ListGroup.Item>
                     <ListGroup.Item
                         as="li"
@@ -42,9 +80,14 @@ export const CreateCampaignS5 = () => {
                         <div className="fw-bold">Step 2</div>
                             Provide detailed information about the campaign
                         </div>
-                        <Badge bg="danger">
+                        { step2 ? 
+                        <Badge bg="success">
+                        Good!
+                        </Badge>
+                        : <Badge bg="danger">
                         Suggestions below
                         </Badge>
+                        }
                     </ListGroup.Item>
                     <ListGroup.Item
                         as="li"
@@ -54,9 +97,14 @@ export const CreateCampaignS5 = () => {
                         <div className="fw-bold">Step 3</div>
                             Location and language
                         </div>
+                        { step3 ? 
                         <Badge bg="success">
                         Good!
                         </Badge>
+                        : <Badge bg="danger">
+                        Suggestions below
+                        </Badge>
+                        }
                     </ListGroup.Item>
                     <ListGroup.Item
                         as="li"
@@ -66,73 +114,27 @@ export const CreateCampaignS5 = () => {
                         <div className="fw-bold">Step 4</div>
                             Additional materials and attachments
                         </div>
+                        { step4 ? 
                         <Badge bg="success">
                         Good!
                         </Badge>
+                        : <Badge bg="danger">
+                        Suggestions below
+                        </Badge>
+                        }
                     </ListGroup.Item>
                     </ListGroup>
                 </Row>
-                <Row className="mt-3">
-                    <hr></hr>
-                    <Button variant="link" href="/create-campaign-step-2">Suggestions for Step 2: Detailed information</Button>
-                    <hr></hr>
-                    <Col>
-                        <Card className="mt-3">
-                            <Card.Body>
-                                <Form>
-                                    <FormGroup>
-                                    <Form.Label>Campaign mission</Form.Label>
-                                    <Form.Control as="textarea" rows={4} type="text" placeholder="Describe your campaign mission here..." />
-                                    <Form.Text className="text-muted">Explainer text about the role of the campaign challenge</Form.Text>
-                                    </FormGroup>
-                                </Form>
-                            </Card.Body>
-                        </Card>
-                    </Col>
-                    <Col>
-                        <Card className="mt-3">
-                            <Card.Body>
-                                <Card.Text>
-                                    Hint 1: This is a hint
-                                </Card.Text>
-                                <Card.Text>
-                                    Hint 2: This is a hint
-                                </Card.Text>
-                                <Card.Text>
-                                    Hint 3: This is a hint
-                                </Card.Text>
-                            </Card.Body>
-                        </Card>
-                    </Col>
-                </Row>
-                <Row className="mt-5">
-                    <Col>
-                        <Card>
-                            <Card.Body>
-                                <Form>
-                                    <FormGroup>
-                                        <Form.Label>Predicted goals</Form.Label>
-                                        <Form.Control type="text" placeholder="Describe your campaign mission here..." />
-                                        <Form.Text className="text-muted">Explainer text about the role of the campaign challenge</Form.Text>
-                                    </FormGroup>
-                                </Form>
-                            </Card.Body>
-                        </Card>
-                    </Col>
-                    <Col>
-                        <Card>
-                            <Card.Body>
-                                <Card.Text>Hint 1: this is a hint</Card.Text>
-                                <Card.Text>Hint 2: this is a hint</Card.Text>
-                                <Card.Text>Hint 3: this is a hint</Card.Text>
-                            </Card.Body>
-                        </Card>
-                    </Col>
-                </Row>
+                <Container>
+                    { step1 ? <></> : <Step1Suggestions/>}
+                    { step2 ? <></> : <Step2Suggestions/>}
+                    {/* { step3 ? <></> : <CreateCampaignS3/>}
+                    { step4 ? <></> : <CreateCampaignS4/>} */}
+                </Container>
                 <Container className="mt-4 mb-5">
                     <Row className="justify-content-md-center">
                         <hr></hr>
-                        <Button size="lg" variant="success" href="create-campaign-step-6">Submit!</Button>
+                        <Button onClick={() => checkSteps()} classId="submit" size="lg" variant="success">Submit!</Button>
                     </Row>
                 </Container>
             </Container>
